Handle error and detail format

+++ src/services/storeService.js
@@ -1,12 +1,19 @@
+const UnauthorizedError = require('../exceptions/UnauthorizedError');
+const NotFoundError = require('../exceptions/NotFoundError');
+const AlreadyExistsError = require('../exceptions/AlreadyExistsError');
 const Store = require('../models/Store');
 const Item = require('../models/Item');
 const User = require("../models/User/User");
 
 const createStore = async (ownerId, storeData) => {
+    const owner = await User.findById(ownerId);
+    if (!owner) {
+        throw new NotFoundError('Owner not found');
+    }
+
     const store = new Store({owner: ownerId, ...storeData});
     await store.save();
 
-    const owner = await User.findById(ownerId);
     owner.store.push(store._id);
     await owner.save();
     return store;
@@ -23,6 +30,10 @@ const updateStore = async (storeId, updateData) => {
 
 const addItemToInventory = async (storeId, itemData) => {
     const store = await Store.findById(storeId);
+    if (!store) {
+        throw new NotFoundError('Store not found');
+    }
+
     const newItem = new Item({store: storeId, ...itemData});
     await newItem.save();
 
@@ -33,15 +44,30 @@ const addItemToInventory = async (storeId, itemData) => {
 };
 
 const updateItemInInventory = async (itemId, updateData) => {
-    return Item.findByIdAndUpdate(itemId, updateData, {new: true});
+    const updatedItem = await Item.findByIdAndUpdate(itemId, updateData, {new: true});
+    if (!updatedItem) {
+        throw new NotFoundError('Item not found');
+    }
+    return updatedItem;
 };
 
+
 const getStoreInfo = async (storeId) => {
-    return Store.findById(storeId).populate('inventory', ['name', 'price']).populate('owner', ['username', 'email']);
+    const store = await Store.findById(storeId)
+        .populate('inventory.item')
+        .populate('owner', ['username', 'email']);
+    if (!store) {
+        throw new NotFoundError('Store not found');
+    }
+    return store;
 };
+
 
 const deleteStore = async (storeId) => {
     const store = await Store.findById(storeId);
+    if (!store) {
+        throw new NotFoundError('Store not found');
+    }
     // console.log(store.owner);
     await Item.deleteMany({store: storeId});
     await Store.findByIdAndDelete(storeId);
@@ -50,6 +76,10 @@ const deleteStore = async (storeId) => {
 
 const deleteItemFromInventory = async (itemId) => {
     const item = await Item.findById(itemId);
+    if (!item) {
+        throw new NotFoundError('Item not found');
+    }
+
     await Store.findByIdAndUpdate(item.owner, {$pull: {inventory: itemId}});
     // Remove the item
     await Item.findByIdAndDelete(itemId);
